[MCZB-2671] Kconfig: allow auto select of the first valid mac interface
Setting the interface as the default value prevents kconfig to select
the first valid mac interface. We can keep this behavior by removing
the default value and moving the interface config to the end of the
choice. This way, kconfig will select the interface symbol only when
the previous ones can't be selected, which means the platform is not
directly supported. The choice can still be overridden in a defconfig
like required by Matter.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>