Issue: WSW-54178 [PHY][RW610n] TX power level rework to support all the Front-End Modules
[Description]
The TX power level function does not support all the Front-End Modules which includes a TX gain

[Root Cause]
The limits are not appropriate in the test range for the phyTxPowerLevel value

[Fix]
The power level range is calculated with signed values.
This permits to check easly if phyTxPowerLevel is included in the power level range, even if a
TX gain is present or not.

[Testing]
txpower tests with the mfg ot-cli commands as the following example :
radio_nxp mfgcmd 1
radio_nxp mfgcmd 15
radio_nxp mfgcmd 16 10
radio_nxp mfgcmd 15

Change-Id: a82abdf2-e889-11ef-b7c8-2bd2d6765137