[MCZB-2792] ASP: switch XCVR to GFSK mode for unmodulated TX
When saving & restoring the DATA_PADDING_CTRL register, the TX_DIG_CTRL
register is not restored, as such no TX can happen after going back to idle
from an unmodulated TX CW.

Furthermore, when storing & restoring the TX_DIG_CTRL, the first 15.4 TX attempt
will render a malformed PSDU on air (no sync symbols, no preamble).

As such, the proposal is to drop the initial DATA_PADDING_CTRL method and rely
on switching the XCVR to GFSK and then (upon idle command) to O-QPSK.

This method though has a drawack: all functions that are accessing the ZLL while
the GFSK LL is enabled will fail (clock is off to the ZLL) with a bus fault.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>