[MATTER-3327][WSW-48924] PHY: time: OVF value is not incremented when new timer is set too close to OVF event
[Description]
When a new timer is set too close to overflow event, the overflow value is not incremented.

Too close => (overflow-new_time <= gPhyTimeMinSetupTime_c)

In that case the overflow timer cannot be reprogrammed and the PHY timer (24 bits) will overflow before the timer triggers.
As at the moment the condition to increment the timer is based on the next event being the overflow event,
if the new timer is slightly inferior to it, the overflow timer will not be triggered and will not be the next event.

Consequences : if a timer was already programmed after the overflow, it will be delayed by 268s

Signed-off-by: Maele Philippot <maele.philippot@nxp.com>
Signed-off-by: Xavier Razavet <xavier.razavet@nxp.com>
Signed-off-by: George Stefan <george.stefan@nxp.com>