[MCZB-2983] ASP: fix AspTelecSendRawData by consuming the TX interrupt "in place"
Ensure that the TX interrupt generated as part of the "spurious TX"
that is trigged by the AspTelecSendRawData command does not affect
the standard packet flow. As such, mask all ZB interrupts before doing
a forced TX state transition and wait for the interrupt to be generated.
Lastly, restore the state of the PHY_CTRL register to its pre-call value.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>