[MCZB-3070] MW30: conn_test: fix TC12 (unmod) tests
On TelecTest(UNMOD) switches to BLE and then GenFsk. While doing that the
ZLL is deactivated (so access to it's memory region yields an exception).
Going back to IDLE we first issue a soft reset by accessing some ZLL register
and therefore we go into HardFault.
The fix is to remove this call and only do TelecTest(IDLE) which resets
the Radio to ZLL mode.